Evolve function - initial concept
Wrote evolve_step and evolve_iterate to run a genetic evolution of the
rooms over time by switching the saddest people. This initial concept
didn't immediately seem to increase happiness off the bat, aka it didn't
work so well. So, will need to rethink this so that the algorithm is
more effective. Maybe it just has a long runtime, but that's still a bad
thing so. Yeah.
Also, I updated lots of the docstrings but not all of them so I will
need to do this when I have time.