Wrote slightly better switches, algorithm iterator
The program now tries every switch and takes the first happier switch it
finds, but if no switches are happier it goes back and does the first
one anyway. This helps decrease stalls but does not eliminate them. As
such, a single run of the algorithm (the evolve_iterate() function)
doesn't always find the global minimum, but may find a local one. So I
implemented an algorithm iteration function (algorithm_iterate()) to run
it repeatedly and return the minimum value found in all those runs. This
works reasonably well, but has an element of brute force to it and is
thus not as elegant as one would like. So, the real task now is to find
out how to eliminate stalls entirely, so one run always finds the global
minimum, or the happiest arrangement.